requirements.txt
# Flask API (backend/src/app.py)
flask>=3.0.0,<4
flask-cors>=4.0.0
SQLAlchemy>=2.0.0,<3
PyMySQL>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.0

# Used by backend/src/views/schemas.py and config.py (optional if you only run Flask)
pydantic>=2.0.0
pydantic-settings>=2.0.0
